Moved utility classes into their own namespace.
