+++ Utilities/GlobalHotkey.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Windows.Forms;
 using System.Windows.Input;
 
-namespace MouseNet.Logophi
+namespace MouseNet.Logophi.Utilities
 {
     public class GlobalHotkey : IDisposable
     {
